Fork Apache Commons Lang 3.4's TypeUtils class
We cannot use it as is, because it has bugs which we will need to fix.
And (at least for now) it is nice to avoid this dependency anyway.

This commit splices in the TypeUtils code from Apache Commons Lang 3.4
with a minimal set of changes which compile and work with SciJava code
style. Subsequent commits will fix bugs as well as clean up various
issues in the code.