UserInterface: make show methods return the viewer
Having show return an object is actually handy for an unintuitive
reason: use from scientific notebooks such as Beaker Notebook.

In Beaker, the result of a cell execution will be rendered beneath the
cell. Making the show methods return an Object enables code such as:

  image = ij.io().open("myAwesomeImage.tif")
  ij.ui().show(image)

As long as we have a UI plugin dedicated to notebooks (i.e., BeakerUI
or NotebookUI), which returns a BufferedImage when showing images, the
above code will do the slick thing and show the image in the notebook.