Logger: add factory methods for ListenableLogger to Logger interface
A client might want to log to a given Logger instance, and at the
same time listen to it's own (and only to it's own) log messages.
The client can achieve this, by deriving a ListenableLogger with the
new methods from the give Logger instance. The client will have to
log and listen to the ListenableLogger. The messages will automatically
be forwarded to the given Logger instance.